clean up and fixes for make

+++ demo-carpet.py
@@ -5,16 +5,23 @@ from mrivis import Collage
 import sys
 
 
+if len(sys.argv)==2:
+    path1 = sys.argv[1]
+    if os.path.isfile(path1):
+        func_carpet = Carpet(epi_path)
+        ax = func_carpet.show()
+        ax.figure.savefig('carpet.png')
+
 # load a func mri volume from the repo
-epi_path = 'mrivis/example_datasets/epi_func.nii'
-from mrivis import Carpet
+#epi_path = 'mrivis/example_datasets/epi_func.nii'
+
 
 # let's build one from the 4D volume
-func_carpet = Carpet(epi_path)
 
 
-ax = func_carpet.show()
+
+
 # ax is the matplotlib Axis containing the Carpet image
 #plt.show() # this command brings up the figure
 
-ax.figure.savefig('carpet.png')
+
